use dispatch/on to separate telemetry writing from config resolution
