+++ ddtrace/internal/telemetry/__init__.py
@@ -4,19 +4,15 @@ This is normally started automatically when ``ddtrace`` is imported. It can be d
 ``DD_INSTRUMENTATION_TELEMETRY_ENABLED`` variable to ``False``.
 """
 
-import typing as t
-
+from ddtrace.internal import core
+from ddtrace.internal.getconfig import _invalid_otel_config
+from ddtrace.internal.getconfig import get_config
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.settings import env
 from ddtrace.internal.settings._agent import config as agent_config
-from ddtrace.internal.settings._core import FLEET_CONFIG
-from ddtrace.internal.settings._core import FLEET_CONFIG_IDS
-from ddtrace.internal.settings._core import LOCAL_CONFIG
 from ddtrace.internal.settings._core import DDConfig
 from ddtrace.internal.settings._otel_remapper import ENV_VAR_MAPPINGS
 from ddtrace.internal.settings._otel_remapper import SUPPORTED_OTEL_ENV_VARS
-from ddtrace.internal.settings._otel_remapper import parse_otel_env
-from ddtrace.internal.settings._supported_configurations import CONFIGURATION_ALIASES
 from ddtrace.internal.settings._supported_configurations import SENSITIVE_CONFIGURATIONS
 from ddtrace.internal.settings.process_tags import process_tags_config
 from ddtrace.internal.telemetry.constants import TELEMETRY_NAMESPACE
@@ -28,94 +24,16 @@ log = get_logger(__name__)
 __all__ = ["telemetry_writer"]
 
 
-def get_config(
-    envs: t.Union[str, list[str]],
-    default: t.Any = None,
-    modifier: t.Optional[t.Callable[[t.Any], t.Any]] = None,
-    otel_env: t.Optional[str] = None,
-    report_telemetry=True,
-) -> t.Any:
-    """Retrieve a configuration value in order of precedence:
-    1. Fleet stable config (highest)
-    2. Datadog env vars
-    3. OpenTelemetry env vars
-    4. Local stable config
-    5. Default value (lowest)
+def _add_conf(*args):
+    telemetry_writer.add_configuration(*args)
 
-    Reports telemetry for every detected configuration source.
-    """
-    if isinstance(envs, str):
-        envs = [envs]
 
-    # Expand with registered aliases of the canonical name (envs[0]) so all
-    # config sources (LOCAL_CONFIG, env, FLEET_CONFIG) honor legacy renames.
-    canonical = envs[0]
-    aliases = CONFIGURATION_ALIASES.get(canonical, [])
-    if aliases:
-        envs += [a for a in aliases if a not in envs]
+def _add_count(*args):
+    telemetry_writer.add_count_metric(*args)
 
-    effective_val = default
-    telemetry_name = envs[0]
-    # Configurations marked ``sensitive: true`` in the registry are excluded from
-    # configuration telemetry, regardless of which source supplies the value.
-    if telemetry_name in SENSITIVE_CONFIGURATIONS:
-        report_telemetry = False
-    if report_telemetry:
-        telemetry_writer.add_configuration(telemetry_name, default, "default")
 
-    for env_name in envs:
-        if env_name in LOCAL_CONFIG:
-            val = LOCAL_CONFIG[env_name]
-            if modifier:
-                val = modifier(val)
-
-            if report_telemetry:
-                telemetry_writer.add_configuration(telemetry_name, val, "local_stable_config")
-            effective_val = val
-            break
-
-    if otel_env is not None and otel_env in env:
-        raw_val, parsed_val = parse_otel_env(otel_env)
-        if parsed_val is not None:
-            val = parsed_val
-            if modifier:
-                val = modifier(val)
-
-            if report_telemetry:
-                # OpenTelemetry configurations always report the raw value
-                telemetry_writer.add_configuration(telemetry_name, raw_val, "otel_env_var")
-            effective_val = val
-        else:
-            _invalid_otel_config(otel_env)
-
-    for env_name in envs:
-        if env_name in env:
-            val = env[env_name]
-            if modifier:
-                val = modifier(val)
-
-            if report_telemetry:
-                telemetry_writer.add_configuration(telemetry_name, val, "env_var")
-                if otel_env is not None and otel_env in env:
-                    _hiding_otel_config(otel_env, env_name)
-            effective_val = val
-            break
-
-    for env_name in envs:
-        if env_name in FLEET_CONFIG:
-            val = FLEET_CONFIG[env_name]
-            config_id = FLEET_CONFIG_IDS.get(env_name)
-            if modifier:
-                val = modifier(val)
-
-            if report_telemetry:
-                telemetry_writer.add_configuration(telemetry_name, val, "fleet_stable_config", config_id)
-                if otel_env is not None and otel_env in env:
-                    _hiding_otel_config(otel_env, env_name)
-            effective_val = val
-            break
-
-    return effective_val
+core.on("telemetry.add_configuration", _add_conf)
+core.on("telemetry.add_count_metric", _add_count)
 
 
 telemetry_enabled = get_config("DD_INSTRUMENTATION_TELEMETRY_ENABLED", True, asbool, report_telemetry=False)
@@ -144,30 +62,21 @@ def report_configuration(config: DDConfig) -> None:
         for p in name.split("."):
             env_val = getattr(env_val, p)
 
-        telemetry_writer.add_configuration(env_name, env_val, config.value_source(env_name), config.config_id)
-
-
-def _invalid_otel_config(otel_env):
-    log.warning(
-        "Setting %s to %s is not supported by ddtrace, this configuration will be ignored.",
-        otel_env,
-        env.get(otel_env, ""),
-    )
-    telemetry_writer.add_count_metric(
-        TELEMETRY_NAMESPACE.TRACERS,
-        "otel.env.invalid",
-        1,
-        (("config_opentelemetry", otel_env.lower()),),
-    )
+        core.dispatch(
+            "telemetry.add_configuration", (env_name, env_val, config.value_source(env_name), config.config_id)
+        )
 
 
 def _unsupported_otel_config(otel_env):
     log.warning("OpenTelemetry configuration %s is not supported by Datadog.", otel_env)
-    telemetry_writer.add_count_metric(
-        TELEMETRY_NAMESPACE.TRACERS,
-        "otel.env.unsupported",
-        1,
-        (("config_opentelemetry", otel_env.lower()),),
+    core.dispatch(
+        "telemetry.add_count_metric",
+        (
+            TELEMETRY_NAMESPACE.TRACERS,
+            "otel.env.unsupported",
+            1,
+            (("config_opentelemetry", otel_env.lower()),),
+        ),
     )
 
 
@@ -185,8 +94,7 @@ def validate_otel_envs():
             otel_value = env.get(otel_env, "none").lower()
             if otel_value != "none":
                 _invalid_otel_config(otel_env)
-            # TODO: Separate from validation
-            telemetry_writer.add_configuration(otel_env, otel_value, "env_var")
+            core.dispatch("telemetry.add_configuration", (otel_env, otel_value, "env_var"))
         elif otel_env == "OTEL_METRICS_EXPORTER":
             # defer validation to validate_and_report_otel_metrics_exporter_enabled
             pass
@@ -202,25 +110,9 @@ def validate_and_report_otel_metrics_exporter_enabled():
         elif otel_value != "otlp":
             _invalid_otel_config("OTEL_METRICS_EXPORTER")
 
-        # Report to configuration telemetry
-        telemetry_writer.add_configuration("OTEL_METRICS_EXPORTER", otel_value, "env_var")
+        core.dispatch("telemetry.add_configuration", ("OTEL_METRICS_EXPORTER", otel_value, "env_var"))
 
     return metrics_exporter_enabled
-
-
-def _hiding_otel_config(otel_env, dd_env):
-    log.debug(
-        "Datadog configuration %s is already set. OpenTelemetry configuration will be ignored: %s=%s",
-        dd_env,
-        otel_env,
-        env[otel_env],
-    )
-    telemetry_writer.add_count_metric(
-        TELEMETRY_NAMESPACE.TRACERS,
-        "otel.env.hiding",
-        1,
-        (("config_opentelemetry", otel_env.lower()), ("config_datadog", dd_env.lower())),
-    )
 
 
 # TODO: Remove this once the telemetry feature is refactored to a better design
